CP-40844 merges download x file actions into single file.
