--- XenModel/Actions/Updates/DownloadSourceAction.cs
@@ -1,55 +1,0 @@
-﻿/* Copyright (c) Cloud Software Group, Inc. 
- * 
- * Redistribution and use in source and binary forms, 
- * with or without modification, are permitted provided 
- * that the following conditions are met: 
- * 
- * *   Redistributions of source code must retain the above 
- *     copyright notice, this list of conditions and the 
- *     following disclaimer. 
- * *   Redistributions in binary form must reproduce the above 
- *     copyright notice, this list of conditions and the 
- *     following disclaimer in the documentation and/or other 
- *     materials provided with the distribution. 
- * 
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND 
- * CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, 
- * INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF 
- * MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE 
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR 
- * CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, 
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, 
- * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR 
- * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS 
- * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, 
- * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING 
- * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE 
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF 
- * SUCH DAMAGE.
- */
-
-using System;
-using System.ComponentModel;
-using System.IO;
-using System.Net;
-using System.Net.Cache;
-using System.Net.NetworkInformation;
-using System.Threading;
-using XenAdmin.Core;
-
-namespace XenAdmin.Actions.Updates
-{
-    public class DownloadSourceAction : DownloadFileAction, IByteProgressAction
-    {        
-        public DownloadSourceAction(string sourceName, Version version, Uri uri, string outputFileName)
-            : base(sourceName + " source", 
-                  uri, 
-                  outputFileName, 
-                  string.Format(Messages.DOWNLOADING_FILE, version.ToString() + " " + outputFileName ), 
-                  false)
-        {
-
-        }
-    }
-}
-
